@@ -1,73 +1,52 @@
 name: twasol
 description: A new Flutter project.
 
-# The following line prevents the package from being accidentally published to
-# pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: 'none'
 
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk:  ">=2.17.1 <3.0.0"
 
-# Dependencies specify other packages that your package needs in order to work.
-# To automatically upgrade your package dependencies to the latest versions
-# consider running `flutter pub upgrade --major-versions`. Alternatively,
-# dependencies can be manually updated by changing the version numbers below to
-# the latest version available on pub.dev. To see which dependencies have newer
-# versions available, run `flutter pub outdated`.
 dependencies:
   flutter:
     sdk: flutter
 
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  smooth_page_indicator: ^1.0.0+2
-  conditional_builder_null_safety: ^0.0.6
-  bloc: ^8.0.2
-  flutter_bloc: ^8.0.1
-  dio: ^4.0.4
-  cubit: ^0.1.2
-  hexcolor: ^2.0.3
-  shared_preferences: ^2.0.11
-  fluttertoast: ^8.0.8
-  carousel_slider: ^4.0.0
-  #FireBase
-  firebase_core: ^1.11.0
-  firebase_auth: ^3.3.5
-  cloud_firestore: ^3.1.6
-  image_picker: ^0.8.4+4
-  firebase_storage: ^10.2.5
-  firebase_messaging: ^11.2.5
-  desktop_window: ^0.4.0
-  responsive_builder: ^0.4.1
-  easy_splash_screen: ^1.0.3
-  intl: ^0.17.0
-  flutter_launcher_icons: ^0.9.2
-  flutter_local_notifications: ^9.4.0
-  google_sign_in: ^5.2.3
-  pull_to_refresh: ^2.0.0
-  permission_handler: ^8.1.6
-  record: ^3.0.3
+  cupertino_icons:
+  smooth_page_indicator:
+  conditional_builder_null_safety:
+  bloc:
+  flutter_bloc:
+  cubit:
+  http:
+  hexcolor:
+  shared_preferences:
+  fluttertoast:
+  carousel_slider:
+  firebase_core:
+  firebase_auth:
+  cloud_firestore:
+  image_picker:
+  firebase_storage:
+  firebase_messaging: ^14.0.3
+  desktop_window:
+  responsive_builder:
+  easy_splash_screen:
+  intl:
+  flutter_launcher_icons:
+#  awesome_notifications:
+  flutter_local_notifications:
+  google_sign_in:
+  pull_to_refresh:
+  permission_handler: ^10.2.0
+  record:
 #  just_audio: ^0.9.20
 #  audio_visualizer: ^0.1.1
-  audioplayers: ^0.20.1
-  timezone: ^0.8.0
-  flutter_audio_waveforms: ^1.1.1+2
-  social_media_recorder: ^0.1.61
-
+  audioplayers:
+  timezone:
+  flutter_audio_waveforms:
+#  social_media_recorder:
 
 
 flutter_icons:
