@@ -1,10 +1,12 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:twasol/Module/register/register_cubit/social_register_states.dart';
+import 'package:twasol/Module/auth/register/register_cubit/social_register_states.dart';
 import 'package:twasol/model/user_model.dart';
 import 'package:twasol/shared/components/constants.dart';
+import 'package:twasol/shared/cubit/social_cubit.dart';
 import 'package:twasol/shared/network/local/cache_helper.dart';
 
 class SocialRegisterCubit extends Cubit<SocialRegisterStates> {
@@ -14,22 +16,25 @@ class SocialRegisterCubit extends Cubit<SocialRegisterStates> {
 
   //Register
 
-  void userRegister({
+  Future userRegister({
+    required BuildContext context,
     required String userName,
     required String? phone,
     required String email,
     required String password,
     required String vPassword,
-  }) {
+  }) async {
     emit(SocialRegisterLoadingStates());
 
-    FirebaseAuth.instance
+    await FirebaseAuth.instance
         .createUserWithEmailAndPassword(
       email: email,
       password: password,
     )
         .then((value) {
+
       userCreate(
+        context: context,
         userName: userName,
         email: email,
         phone: phone,
@@ -37,11 +42,13 @@ class SocialRegisterCubit extends Cubit<SocialRegisterStates> {
       );
       CacheHelper.saveData(key: 'uId', value: value.user!.uid).then((value) {
         value
-            ? print('uId saved in Cache Memory Successfully')
-            : print('Failed save uId in Cache Memory');
+            ? debugPrint('uId saved in Cache Memory Successfully')
+            : debugPrint('Failed save uId in Cache Memory');
       });
       uId = value.user!.uid;
       token = value.credential!.token;
+
+      emit(SocialRegisterSuccessStates());
     }).catchError((error) {
       emit(SocialRegisterErrorStates(error.toString()));
     });
@@ -49,31 +56,36 @@ class SocialRegisterCubit extends Cubit<SocialRegisterStates> {
 
   // Cloud FireStore Create
 
-  void userCreate({
+  Future<void> userCreate({
+    required BuildContext context,
     required String userName,
     required String email,
     required String? phone,
     required String uId,
-  }) {
+  }) async {
     emit(SocialRegisterLoadingStates());
 
     UserModel userModel = UserModel(
-      userName: userName ,
+      userName: userName,
       email: email,
       phone: phone!,
-      uId:  uId,
-      imgCover: 'https://i.kym-cdn.com/entries/icons/original/000/034/213/cover2.jpg',
-      image: 'https://us.123rf.com/450wm/fizkes/fizkes2007/fizkes200701793/152407909-profile-picture-of-smiling-young-caucasian-man-in-glasses-show-optimism-positive-and-motivation-head.jpg?ver=6',
+      uId: uId,
+      imgCover:
+          'https://i.kym-cdn.com/entries/icons/original/000/034/213/cover2.jpg',
+      image:
+          'https://us.123rf.com/450wm/fizkes/fizkes2007/fizkes200701793/152407909-profile-picture-of-smiling-young-caucasian-man-in-glasses-show-optimism-positive-and-motivation-head.jpg?ver=6',
       bio: 'bio..',
       isEmailVerified: false,
     );
 
-    FirebaseFirestore.instance
+    await FirebaseFirestore.instance
         .collection('users')
         .doc(uId)
         .set(userModel.toMap())
         .then((value) {
-          FirebaseMessaging.instance.subscribeToTopic('users');
+
+      FirebaseMessaging.instance.subscribeToTopic('users');
+      SocialCubit.get(context).getFcmToken();
       emit(SocialCreateUserSuccessStates());
     }).catchError((error) {
       emit(SocialCreateUserErrorStates(error.toString()));
