@@ -1,26 +1,35 @@
-import 'package:audioplayers/notifications.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:twasol/Module/auth/login/cubit/social_login_cubit.dart';
+import 'package:twasol/Module/auth/register/register_cubit/social_register_cubit.dart';
 import 'package:twasol/layout/splash/splash_screen.dart';
 import 'package:twasol/shared/bloc_observer.dart';
 import 'package:twasol/shared/components/constants.dart';
-import 'package:twasol/shared/components/notifications_service.dart';
 import 'package:twasol/shared/components/permissions.dart';
 import 'package:twasol/shared/cubit/social_cubit.dart';
 import 'package:twasol/shared/network/local/cache_helper.dart';
 import 'shared/styles/themes.dart';
 
-
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
   await CacheHelper.init();
-  NotificationServices().initNotification();
-  if(permissionsGranted == null || permissionsGranted == false){
-    await PermissionHandler.appPermission();
-  }
+  // NotificationServices().initNotification();
+  // AwesomeNotifications().initialize(
+  //   'resource://drawable/res_app_icon',
+  //   [
+  //     NotificationChannel(
+  //       channelKey: 'basic_channel',
+  //       channelName: 'Basic notifications',
+  //       channelDescription: 'Notification channel for basic tests',
+  //       defaultColor: Colors.blue,
+  //       ledColor: Colors.white,
+  //     ),
+  //   ],
+  // );
+
   //when the app is opened
   FirebaseMessaging.onMessage.listen((event) {});
   // when click on notification to open app
@@ -28,19 +37,13 @@ void main() async {
   // background notification
   // FirebaseMessaging.onBackgroundMessage(firebaseMessagingBackgroundHandler);
 
-
-  var firebaseMessagingToken = await FirebaseMessaging.instance.getToken();
-  print('=================token=============================');
-  print('firebaseMessagingToken $firebaseMessagingToken');
-
   // bool onBoarding = false;
   uId = CacheHelper.getData(key: 'uId');
   permissionsGranted = CacheHelper.getData(key: 'permissionsGranted');
 
-
   BlocOverrides.runZoned(
     () {
-      runApp(SplashScreen());
+      runApp(const SplashScreen());
     },
     blocObserver: MyBlocObserver(),
   );
@@ -49,9 +52,11 @@ void main() async {
 class MyApp extends StatelessWidget {
   final Widget? startWidget;
 
-  MyApp({
+  const MyApp({
+    super.key,
     this.startWidget,
   });
+
   @override
   Widget build(BuildContext context) {
     return MultiBlocProvider(
@@ -59,12 +64,16 @@ class MyApp extends StatelessWidget {
         BlocProvider(
           create: (context) => SocialCubit()
             ..getDataUser()
-            ..getPosts(),
-        )
+            ..getPosts()
+            ..getFcmToken()
+            ..getNotifications(),
+        ),
+        BlocProvider(create: (context) => SocialLoginCubit()),
+        BlocProvider(create: (context) => SocialRegisterCubit()),
       ],
       child: MaterialApp(
         debugShowCheckedModeBanner: false,
-        title: 'Social App',
+        title: 'Twasol',
         theme: lightMode,
         darkTheme: darkMode,
         themeMode: ThemeMode.light,
@@ -73,4 +82,3 @@ class MyApp extends StatelessWidget {
     );
   }
 }
-
