@@ -73,7 +73,6 @@ class SocialCreatePostErrorStates extends SocialStates {
   final String error;
 
   SocialCreatePostErrorStates(this.error);
-
 }
 
 class SocialRemovePostImageStates extends SocialStates {}
@@ -117,10 +116,15 @@ class SocialCommentsPostsErrorStates extends SocialStates {
 
   SocialCommentsPostsErrorStates(this.error);
 }
+
 class SocialLikedLoadingStates extends SocialStates {}
+
 class SocialLikedSuccessStates extends SocialStates {}
+
 class SocialLikedByMeCheckedLoadingState extends SocialStates {}
+
 class SocialLikedByMeCheckedSuccessState extends SocialStates {}
+
 class SocialLikedByMeCheckedErrorState extends SocialStates {
   final String error;
 
@@ -128,23 +132,20 @@ class SocialLikedByMeCheckedErrorState extends SocialStates {
 }
 
 class SocialDisLikePostSuccessState extends SocialStates {}
+
 class SocialDisLikePostErrorState extends SocialStates {
-  final String error ;
+  final String error;
 
   SocialDisLikePostErrorState(this.error);
 }
 
-
 class SocialGetLikesCountLoadingStates extends SocialStates {}
+
 class SocialGetLikesCountSuccessStates extends SocialStates {}
-
-
-
 
 class SocialGetCommentsLoadingStates extends SocialStates {}
 
 class SocialGetCommentsSuccessStates extends SocialStates {}
-
 
 class SocialSendMessageSuccessStates extends SocialStates {}
 
@@ -152,7 +153,17 @@ class SocialSendMessageErrorStates extends SocialStates {}
 
 class SocialGetMessagesSuccessStates extends SocialStates {}
 
+class SocialSaveFcmTokenSuccessStates extends SocialStates {}
+
 class SocialGetMessagesLoadingStates extends SocialStates {}
+
+class SocialGetFcmTokenSuccessStates extends SocialStates {}
+
+class SocialGetFcmTokenErrorStates extends SocialStates {
+  final String error;
+
+  SocialGetFcmTokenErrorStates(this.error);
+}
 
 class SocialDeletePostSuccessStates extends SocialStates {}
 
@@ -161,37 +172,60 @@ class SocialDeletePostErrorStates extends SocialStates {
 
   SocialDeletePostErrorStates(this.error);
 }
+
 class SocialSignOutLoadingStates extends SocialStates {}
+
 class SocialSignOutSuccessStates extends SocialStates {}
+
 class SocialSignOutErrorStates extends SocialStates {
   final String error;
 
   SocialSignOutErrorStates(this.error);
-
 }
+
 class SocialSearchUserNameLoadingStates extends SocialStates {}
+
 class SocialSearchUserNameSuccessStates extends SocialStates {}
+
 class SocialSearchUserNameErrorStates extends SocialStates {
   final String error;
 
   SocialSearchUserNameErrorStates(this.error);
 }
+
 class SocialOpenLikeSheetStates extends SocialStates {}
+
+class PickChatImageSuccessState extends SocialStates {}
+
+class PickChatImageErrorState extends SocialStates {
+  final String error;
+
+  PickChatImageErrorState(this.error);
+}
+
 class SocialOpenCommentSheetStates extends SocialStates {}
 
 class SocialVoiceRecordOn extends SocialStates {}
+
 class SocialVoiceRecordOff extends SocialStates {}
 
-class SocialVoiceRecorderWithBottomSheetWithBottomSheetLoadingStates extends SocialStates {}
+class SocialVoiceRecorderWithBottomSheetWithBottomSheetLoadingStates
+    extends SocialStates {}
+
 class SocialVoiceRecorderWithBottomSheetSuccessStates extends SocialStates {}
+
 class SocialVoiceRecorderWithBottomSheetErrorStates extends SocialStates {
   final String error;
+
   SocialVoiceRecorderWithBottomSheetErrorStates(this.error);
 }
+
 class SocialCheckRecordingStates extends SocialStates {}
 
 class SocialUploadVoiceRecordLoadingStates extends SocialStates {}
+
 class SocialUploadVoiceRecordSuccessStates extends SocialStates {}
+
 class SocialUploadVoiceRecordErrorStates extends SocialStates {
   final String error;
 
@@ -201,7 +235,11 @@ class SocialUploadVoiceRecordErrorStates extends SocialStates {
 class SocialAudioVoiceLoadingStates extends SocialStates {}
 
 class SocialAudioPlayStates extends SocialStates {}
+
 class SocialAudioPauseStates extends SocialStates {}
+
 class SocialAudioResumeStates extends SocialStates {}
+
 class SocialAudioStopStates extends SocialStates {}
+
 class SocialOnAudioPositionChangedStates extends SocialStates {}
